producers/requirements.txt
confluent-kafka==2.3.0
requests==2.31.0
python-dotenv==1.0.0
colorlog==6.8.0